Update conversation.py
TextPart Class Modification:
Added a new boolean field is_context. This field is initialized to False by default. When constructing a TextPart for a contextual piece of text, set it to True.

Message.encode Method Modification:
When a TextPart is processed, check the value of the is_context field. If it is True, encode the text inside with special <|context_start|> and <|context_end|> tokens, this is to inform the LLM that this is indeed context.
If is_context is False, the text part will be handled normally without the context wrapping tokens.